refactor gossip into pkg
